+++ chain/chain/Cargo.toml
@@ -28,6 +28,9 @@ near-metrics = { path = "../../core/metrics" }
 near-pool = { path = "../pool" }
 near-logger-utils = {path = "../../test-utils/logger"}
 
+[dev-dependencies]
+near-store-validator = { path = "../../test-utils/store-validator" }
+
 [features]
 # if enabled, we assert in most situations that are impossible unless some byzantine behavior is observed.
 byzantine_asserts = []
