+++ core/store/Cargo.toml
@@ -18,7 +18,6 @@ rand = "0.7.2"
 
 borsh = "0.6.1"
 
-near-chain-configs = { path = "../chain-configs" }
 near-crypto = { path = "../crypto" }
 near-primitives = { path = "../primitives" }
